requirements.txt
python-dotenv==1.0.1
google-generativeai==0.8.3
google-api-python-client==2.158.0
